Updated credits and removed the jump extend move
I updated the credits to reflect the added music tracks and models from the last commit.

I also removed the buggy jump extend move as it was not working as intended. The enemy combo system was removed due to high ranks not being near impossible now.

Did a quick rewrite to the restart level button so it isn't hardcoded per level anymore