Added various kinds of moving platforms
Commit title is exactly what changed here. 4 types of moving platforms, 3 of which are just the platform moving on the X, Y, and Z axis and a platform that moves in a circle.

Also added the glass from that one game in Squid Game where it randomly can break when you step on it