Expanded Level 1 and bug fixes
I expanded Level 1 to be longer and made it so you can't die in the shop area anymore.

Also gave the shoot enemy type a major improvement as they actually now rotate in the X,Y, and Z axis so their aim is a lot better.

Added a wall jump tutorial clip and a clip that plays when you can't afford a shop item.